# backend-api-handler/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Travel Agency API",
    "description": "Sprint 2 — Auth, Tours, Bookings, and Tour Booking Lifecycle Management. Includes JWT-based authentication, tour browsing, booking lifecycle (BOOKED → CONFIRMED → STARTED → FINISHED / CANCELED), document upload & verification, and travel-agent dedicated endpoints.",
    "contact": {
      "name": "Travel Agency Team",
      "email": "support@travelagency.com"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "Bearer": []
    }
  ],
...
  "paths": {
...
    "/api/v1/tours/{id}/reviews": {
      "get": {
        "tags": [
          "Tours"
        ],
        "summary": "Get reviews for a specific tour",
        "operationId": "getReviews",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 4
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "RATING_DESC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated reviews",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "Tour not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReviewDTO": {
        "type": "object",
        "properties": {
          "authorName": {
            "type": "string"
          },
          "authorImageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "rate": {
            "type": "integer",
            "format": "int32"
          },
          "reviewContent": {
            "type": "string"
          }
        }
      },
      "ReviewListResponseDTO": {
        "type": "object",
        "properties": {
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewDTO"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Paste your JWT access token here",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}